+++ packages/jsActions/mobile-resources-native/typings/mx.d.ts
@@ -1,0 +1,5 @@
+declare namespace mx {
+    interface MxInterface {
+        readFileBlob: (filePath: string) => Promise<string>;
+    }
+}