+++ packages/pluggableWidgets/background-image-native/typings/BackgroundImageProps.d.ts
@@ -30,5 +30,5 @@ export interface BackgroundImagePreviewProps {
     image: { type: "static"; imageUrl: string; } | { type: "dynamic"; entity: string; } | null;
     resizeMode: ResizeModeEnum;
     opacity: number | null;
-    content: { widgetCount: number; renderer: ComponentType<{ caption?: string }> };
+    content: { widgetCount: number; renderer: ComponentType<{ children: ReactNode; caption?: string }> };
 }