+++ packages/pluggableWidgets/toggle-buttons-native/src/ToggleButtons.tsx
@@ -30,7 +30,7 @@ export class ToggleButtons extends Component<Props> {
                     selectedIndex={selectedIndex}
                     enabled={enabled}
                     onTabPress={this.onChangeHandler}
-                    borderRadius={this.styles.container.borderRadius}
+                    borderRadius={Number(this.styles.container.borderRadius)}
                     tabStyle={this.styles.button}
                     tabTextStyle={this.styles.text}
                     activeTabStyle={this.styles.activeButton}