+++ packages/pluggableWidgets/gallery-native/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+-   We've resolved an issue where the loading indicator was triggered when pulling down the list, even in the absence of a pull-down event.
+
+## [1.0.2] - 2023-5-24
+
+### Fixed
+
+-   We've updated react-native-device-info library.
+
 ## [1.0.0] - 2022-10-26
 
 ### Added