+++ packages/pluggableWidgets/image-native/CHANGELOG.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [2.0.0] - 2023-01-24
+
+## BREAKING
+
+-   We made the widget compatible with React Native 0.72.7
+
+### Changed
+
+-   Updated dependency `react-native-vector-icons` from version 9.1.0 to 10.0.3
+
+## [1.1.0] - 2023-10-11
+
+### Added
+
+-   Added accessibility.
+
 ## [1.0.2] - 2022-9-20
 
 ### Changed