+++ packages/pluggableWidgets/video-player-native/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [5.0.0] - 2023-01-24
+
+### Changed
+
+-   Updated dependency `react-native-vector-icons` from version 9.1.0 to 10.0.3
+
+## BREAKING
+
+-   We have upgraded react-native-system-navigation-bar to version 2.6.3
+
 ## [4.0.0] - 2023-3-28
 
 ## BREAKING